// src/interfaces/routes.interface.ts
export enum MainRoutes {
    LOGGER = '/api/log',
    GET_ALL_SYSTEM_LOG = "/api/system-log",
    GET_FAULTS = "/api/faults",
    COMMANDS = "/api/commands",
    SERVER_POLYGONS = '/api/polygons',
    SERVER_MAPS = '/api/maps',
    SERVER_BIT_REPORT_FILES = '/api/bit-report-files',
    LOGIN = '/api/login',
    ALERTS = '/api/alerts',
    ESDUMP = '/api/esdump',
    HEALTH = '/health',
    VERSION = '/version',
    EXPLANATION = '/api/explanation',
    SYSTEM_INFO_LIST = '/api/system-info-list',
    DASHBOARD = '/api/dashboard',
    SYSTEM_PRODUCTS = '/api/system-products',
    FILE_BROWSER_TREE = '/file-tree',
    MAGIC_CHAT_INFO = '/api/magic-chat-info',
    CALIBRATION = '/api/calibration',
}

export enum SubRoutes {
    MAIN = '/',
    RUN_COMMAND = '/run',
    BIT = '/bit',
    TREE = '/tree',
    SYSTEM_INFO = '/system-info',
    DYNAMIC_DIAGRAM = '/dynamic-diagram',
    VALIDATE_TOKEN = 'validateToken',
    BIT_REPORT_LOG_CLICK = '/bit-report-log-click',
    CALIBRATION_LOAD = '/calibration-load',
    CALIBRATION_DELETE = '/calibration-delete',
    BIT_REPORT_TREE_MAP_CLICK = '/bit-report-tre-map-click',
    MULTI_EXPORT = '/multiexport',
    MULTI_IMPORT = '/multiimport',
    MULTI_DELETE = '/multidelete',
    IS_ALIVE = '/isalive',
    IS_READY = '/isready',
    IS_STARTED = '/isstarted',
    METRICS = '/metrics',
    VERSION = '/version',
    SERVICEABILITY = '/serviceability',
    FAULT = '/fault',
    EVENT = '/event',
    UPDATE_FAVORITE = '/updateFavorite',
    SYMON_DATA_MAPPER = '/symon-data-mapper',
    PRODUCTS = '/products',
}